Se agregó Sweetalert para limpiar carrito con promesa

+++ js/cart.js
@@ -1,31 +1,31 @@
-const cartItems = JSON.parse(localStorage.getItem("shoppingCart")) || [];
+const cartItems = JSON.parse(localStorage.getItem("shoppingCart")) || []
 
 function groupAndSumItems(cartItems) {
-    const groupedItems = {};
+    const groupedItems = {}
 
     cartItems.forEach(item => {
         if (!groupedItems[item.id]) {
-            groupedItems[item.id] = { ...item };
+            groupedItems[item.id] = { ...item }
         } else {
-            groupedItems[item.id].quantity += item.quantity;
+            groupedItems[item.id].quantity += item.quantity
         }
     });
 
-    return Object.values(groupedItems);
+    return Object.values(groupedItems)
 }
 
 function renderCart(cartItems) {
-    const cartSection = document.getElementById("card");
-    cartSection.innerHTML = "";
+    const cartSection = document.getElementById("card")
+    cartSection.innerHTML = ""
 
     let subtotal = 0;
 
     cartItems.forEach(item => {
-        const itemTotalPrice = item.price * item.quantity;
-        subtotal += itemTotalPrice;
+        const itemTotalPrice = item.price * item.quantity
+        subtotal += itemTotalPrice
 
-        const cartItemDiv = document.createElement("div");
-        cartItemDiv.classList.add("cart-item");
+        const cartItemDiv = document.createElement("div")
+        cartItemDiv.classList.add("cart-item")
 
         cartItemDiv.innerHTML = `
             <img src="${item.imgUrl}" alt="${item.itemName}" class="cart-item-image">
@@ -38,38 +38,58 @@ function renderCart(cartItems) {
             </div>
         `;
 
-        const deleteButton = cartItemDiv.querySelector(`#delete${item.id}FromCart`);
-        deleteButton.addEventListener("click", () => deleteFromCart(item));
+        const deleteButton = cartItemDiv.querySelector(`#delete${item.id}FromCart`)
+        deleteButton.addEventListener("click", () => deleteFromCart(item))
 
-        cartSection.appendChild(cartItemDiv);
+        cartSection.appendChild(cartItemDiv)
     });
 
-    const total = subtotal * 1.21;
+    const total = subtotal * 1.21
 
-    const cartTotalDiv = document.createElement("div");
-    cartTotalDiv.classList.add("cart-total");
+    const cartTotalDiv = document.createElement("div")
+    cartTotalDiv.classList.add("cart-total")
     cartTotalDiv.innerHTML = `
         <p class="subtotal">Subtotal: $${subtotal.toFixed(2)}</p>
         <p class="total">Total (IVA incluido): $${total.toFixed(2)}</p>
         <button id="clearCartButton" class="btn btn-info">Limpiar carrito de compras</button>
     `;
 
-    cartSection.appendChild(cartTotalDiv);
+    cartSection.appendChild(cartTotalDiv)
 }
 
-const groupedCartItems = groupAndSumItems(cartItems);
-renderCart(groupedCartItems);
+const groupedCartItems = groupAndSumItems(cartItems)
+renderCart(groupedCartItems)
 
 function deleteFromCart(item) {
-    const updatedCartItems = cartItems.filter(cartItem => cartItem.id !== item.id);
-    localStorage.setItem("shoppingCart", JSON.stringify(updatedCartItems));
-    location.reload();
+    const updatedCartItems = cartItems.filter(cartItem => cartItem.id !== item.id)
+    localStorage.setItem("shoppingCart", JSON.stringify(updatedCartItems))
+    location.reload()
 }
 
 const clearCartButton = document.getElementById("clearCartButton")
 clearCartButton.addEventListener("click", () => clearCart())
 
 function clearCart() {
-    localStorage.clear()
-    location.reload()
+        Swal.fire({
+        text: '¿Está seguro que desea limpiar el carrito?',
+        icon: 'alert',
+        showCancelButton: true,
+        confirmButtonText: 'Limpiar carrito', // true
+        cancelButtonText: 'Cancelar' // false
+    })
+    .then( resultado => {
+        if (resultado.isConfirmed) {
+            Swal.fire({
+                icon: 'success',
+                text: 'Se ha limpiado el carrito',
+                showCancelButton: false,
+                showConfirmButton: false
+            })  
+            localStorage.clear()
+            setTimeout(()=>{
+                location.reload()
+            }, 2500)
+        }
+    })
 }
+
